Bug fix setting opacity

+++ core/utils/tools_gw.py
@@ -1052,7 +1052,7 @@ def set_style_mapzones():
                 for id in mapzone['values']:
                     # initialize the default symbol for this geometry type
                     symbol = QgsSymbol.defaultSymbol(lyr.geometryType())
-                    symbol.setOpacity(float(mapzone['opacity']))
+                    symbol.setOpacity(int(mapzone['opacity']))
 
                     # Setting simp
                     R = random.randint(0, 255)
